refactor: move Step 1.4 tests to separate file per CONTRIBUTING.md
Separate 5 new tests from inline #[cfg(test)] module into
src/llm/model_discovery_tests.rs. Historical tests remain in place.

Changes:
- Create model_discovery_tests.rs with 5 Step 1.4 tests
- Remove those tests from inline module in model_discovery.rs
- Add #[cfg(test)] mod model_discovery_tests to mod.rs
- Make ModelDiscovery fields pub(crate) for cross-module test access
- Remove unused InputType import from inline tests

Source: CI
Type: refactor
